requirements.txt
flask==3.0.0
flask-bcrypt==1.0.1
flask-sqlalchemy==3.1.1
python-docx==1.1.0
docx2pdf==0.1.8
reportlab==4.2.2
python-dotenv==1.0.1
